% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill missing water discharge data using linear interpolation}
\usage{
fill_gaps(hdata, autocorr = 0.7, nobserv = NULL, expand = TRUE)
}
\arguments{
\item{hdata}{Water discharge series. Data frame containing 4 columns: YYYY, MM, DD and water level}

\item{autocorr}{Autocorrelation value that defines possible length of the period}

\item{nobserv}{A number of obsrvations to fill}

\item{expand}{Should the algorithm insert missing dates?}
}
\value{
filled discharge values
}
\description{
Fill missing water discharge data using linear interpolation
}
