% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{get_baseflow}
\alias{get_baseflow}
\title{Extracts baseflow for discharge}
\usage{
get_baseflow(
  Q,
  alpha = 0.925,
  padding = 30,
  passes = 3,
  method = "Lyne-Hollick"
)
}
\arguments{
\item{Q}{daily discharge vector}

\item{alpha}{filtering parameter. Defaults to 0.925}

\item{padding}{number of elements padded at the beginning and ending of discharge vector to reduce boundary effects}

\item{passes}{number of filtering iterations. The first iteration is forward, second is backward, third is forward and so on. Defaults to 3}

\item{method}{baseflow filtering method. Currently only `method = LYNE` is supported which corresponds to Lyne-Hollick (1979)}
}
\value{
baseflow vector
}
\description{
Extracts baseflow for discharge
}
