% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_parameters}
\alias{get_parameters}
\title{Gete hydrograph parameters list}
\usage{
get_parameters()
}
\value{
data.frame of parameters
}
\description{
Gete hydrograph parameters list
}
\examples{
grwat::get_parameters()
}
