% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{get_separation_params}
\alias{get_separation_params}
\title{Get default separation parameters for selected region}
\usage{
get_separation_params(type = "Midplain")
}
\arguments{
\item{region}{Name of the region}
}
\value{

}
\description{
Get default separation parameters for selected region
}
