% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_variables}
\alias{get_variables}
\title{Get hydrograph parameters list}
\usage{
get_variables()
}
\value{
data.frame of parameters
}
\description{
Get hydrograph parameters list
}
\examples{
grwat::get_variables()
}
