% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_baseflow}
\alias{gr_baseflow}
\title{Extract baseflow from hydrological series}
\usage{
gr_baseflow(
  Q,
  a = 0.925,
  k = 0.975,
  C = 0.05,
  aq = -0.5,
  passes = 3,
  padding = 30,
  method = "lynehollick"
)
}
\arguments{
\item{Q}{daily discharge vector}

\item{passes}{number of filtering iterations. The first iteration is forward, second is backward, third is forward and so on. Defaults to 3}

\item{padding}{number of elements padded at the beginning and ending of discharge vector to reduce boundary effects}

\item{method}{baseflow filtering method. Available methods are "maxwell", "boughton", "jakeman", "lynehollick", "chapman". Default is "lynehollick", which corresponds to Lyne-Hollick (1979) hydrograph separation method.}

\item{alpha}{filtering parameter. Defaults to 0.925}
}
\value{
baseflow vector
}
\description{
Extract baseflow from hydrological series
}
