% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_check_data}
\alias{gr_check_data}
\title{Check the correctness of data frame for separating}
\usage{
gr_check_data(df)
}
\arguments{
\item{df}{\code{data.frame} with four columns: date, discharge, temperature, precipitation.}
}
\value{
stops execution if \code{df} contains the wrong number of columns, or the columns have the wrong types, or the data in columns is incorrect (e.g. discharge or precipitation are negative).
}
\description{
Check the correctness of data frame for separating
}
\examples{
library(grwat)

# example Spas-Zagorye data is included with grwat package
data(spas)
print(spas)

gr_check_data(spas)

# raw Spas-Zagorye data represents date components
# in columns and does not contain meteorologgical variables
path = system.file("extdata", "spas-zagorye.txt", 
                   package = "grwat")

hdata_raw = readr::read_delim(path, 
                       col_names = c('d', 'm', 'y', 'q'), 
                       col_types = 'iiid', delim = ' ')

print(hdata_raw)

gr_check_data(hdata_raw)



}
