% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{gr_fill_gaps}
\alias{gr_fill_gaps}
\title{Fill missing water discharge data using linear interpolation}
\usage{
gr_fill_gaps(hdata, autocorr = 0.7, nobserv = NULL)
}
\arguments{
\item{hdata}{Water discharge series. Data frame containing 4 columns: year (YYYY), montth (MM), day (DD) and level (water level)}

\item{autocorr}{Autocorrelation value that defines possible length of the period. Defaults to 0.7.  If \code{nobserv} parameter is set, then this parameter is ignored. If both parameters are \code{NULL}, then all gaps are filled disregard of their lengths (not recommended).}

\item{nobserv}{Maximum number of contiguous observations that can be filled. Defaults to \code{NULL}. If this parameter is set by the user, then \code{autocorr} parameter is ignored. If both parameters are \code{NULL}, then all gaps are filled disregard of their lengths (not recommended).}

\item{expand}{Should the algorithm insert missing dates?}

\item{dates}{}

\item{order}{}
}
\value{
filled discharge values
}
\description{
Fill missing water discharge data using linear interpolation
}
