% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_from_pardf}
\alias{gr_from_pardf}
\title{Convert data frame of parameters to list}
\usage{
gr_from_pardf(pardf)
}
\arguments{
\item{params}{\code{tibble} of hydrograph separation parameters as returned by \code{\link[=gr_to_pardf]{gr_to_pardf()}}.}
}
\value{
\code{list} of \code{list}s of hydrograph separation parameters as returned in \code{params} attribute by \code{\link[=gr_separate]{gr_separate()}}  with \code{debug = TRUE}.
}
\description{
Convert data frame of parameters to list
}
