% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_help_params}
\alias{gr_help_params}
\title{Get the information about parameters used to separate the hydrograph}
\usage{
gr_help_params()
}
\value{
\code{data.frame} with description of hydrograph separation parameters that are used in \code{\link[=gr_separate]{gr_separate()}} .
}
\description{
Get the information about parameters used to separate the hydrograph
}
\examples{
library(grwat)

gr_help_params()
}
