% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gr_help_vars}
\alias{gr_help_vars}
\title{Get hydrograph parameters list}
\usage{
gr_help_vars()
}
\value{
data.frame of parameters
}
\description{
Get hydrograph parameters list
}
\examples{
grwat::gr_help_vars()
}
