% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{gr_join_interim}
\alias{gr_join_interim}
\title{Join ERA-INTERIM data to hydrological level series}
\usage{
gr_join_interim(hdata, rean, buffer)
}
\arguments{
\item{hdata}{Water discharge series. Data frame containing 4 columns: YYYY, MM, DD and water level}

\item{rean}{List. Contains three components: days series, values and location points}

\item{buffer}{Region to select reanalysis data}
}
\value{
List of two components: water level table enriched with reanalysis data
  and selected points
}
\description{
Join ERA-INTERIM data to hydrological level series
}
\examples{
\dontrun{
grwat::process(hdata, rean, buffer)
}
}
