% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{gr_join_rean}
\alias{gr_join_rean}
\title{Join grwat reanalysis to hydrological level series}
\usage{
gr_join_rean(hdata, rean, buffer)
}
\arguments{
\item{hdata}{data frame containing 2 columns: date and discharge level}

\item{rean}{List. Contains three components: days series, values and location points}

\item{buffer}{Region to select reanalysis data}
}
\value{
List of two components: water level table enriched with reanalysis data
  and selected points
}
\description{
Join grwat reanalysis to hydrological level series
}
\examples{
\dontrun{
grwat::process(hdata, rean, buffer)
}
}
