% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_ridge}
\alias{gr_plot_ridge}
\title{A convenient wrapper around \code{ggridges::geom_ridgeline()} to visualize multiple river hydrogrphs at once}
\usage{
gr_plot_ridge(df, years, pal = 1, scale = 0.01, alpha = 0.8, locale = "EN")
}
\arguments{
\item{df}{A data.frame with date (1st) and discharge (2nd) columns.}

\item{years}{Integer vector of years to be plotted.}

\item{pal}{Numeric or character string. Color palette identifier passed to \code{ggplot2::scale_fill_brewer()}}

\item{scale}{Numeric. Scale factor passed to \code{ggridges::geom_ridgeline()}. Defaults to 0.01.}

\item{alpha}{Numeric. Opacity value of the ridgeline plot. Defaults to 0.8}
}
\value{
A \code{ggplot2} object
}
\description{
A convenient wrapper around \code{ggridges::geom_ridgeline()} to visualize multiple river hydrogrphs at once
}
