% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_tests}
\alias{gr_plot_tests}
\title{Plot change year density}
\usage{
gr_plot_tests(tests, locale = "EN")
}
\arguments{
\item{tests}{List of tests generated by \code{test_variables()}}

\item{locale}{String locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{
plots density plot for change year
}
\description{
Plot change year density
}
