% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{gr_read_era5_daily}
\alias{gr_read_era5_daily}
\title{Read ERA5 daily reanalysis data}
\usage{
gr_read_era5_daily(file_prec, file_temp)
}
\arguments{
\item{file_prec}{Character string path to daily sum precipitation NetCDF file.}

\item{file_temp}{Character string path to daily mean temperature NetCDF file.}
}
\value{
\code{list} containing time series, precipitation series, temperature series
and spatial points (sf)
}
\description{
The function reads meteorological variables (temperature and precipitation) from \href{https://cds.climate.copernicus.eu/datasets/derived-era5-single-levels-daily-statistics?tab=download}{ERA 5 post-processed daily statistics on single levels} for using with \code{\link[=gr_join_rean]{gr_join_rean()}}.
}
\examples{
if (require("sf") && require("ncdf4")) {
  
  library(grwat)
  
  prec_path = system.file("extdata", "era5_daily_prec.nc",
                          package = "grwat")
  
  temp_path = system.file("extdata", "era5_daily_temp.nc",
                          package = "grwat")
  
  # read reanalysis data
  rean = gr_read_era5_daily(prec_path, temp_path) 
  
  str(rean)
}
}
