% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{gr_read_sep}
\alias{gr_read_sep}
\title{Read file with hydrograph separation}
\usage{
gr_read_sep(file_sep)
}
\arguments{
\item{file_sep}{Path to a file with hydrograph separation data}
}
\value{
data.frame
}
\description{
Read file with hydrograph separation
}
