% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{gr_report}
\alias{gr_report}
\title{Generates HTML report based on separation and variables}
\usage{
gr_report(
  sep,
  vars,
  output = "Report.html",
  year = NULL,
  exclude = NULL,
  locale = "EN"
)
}
\arguments{
\item{sep}{}

\item{vars}{}

\item{output}{}

\item{year}{}

\item{exclude}{}

\item{locale}{}
}
\value{

}
\description{
Generates HTML report based on separation and variables
}
