% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{gr_report_basins}
\alias{gr_report_basins}
\title{Generate reports with detailed hydrograph analysis}
\usage{
gr_report_basins(
  wd,
  year = NULL,
  exclude = NULL,
  map = FALSE,
  locale = "EN",
  pdf = FALSE
)
}
\arguments{
\item{wd}{character string. A path to a working directory with specified structure}

\item{year}{integer. A fixed value for separation year}

\item{exclude}{integer vector. A vector of years to be excluded from tests}

\item{map}{TRUE/FALSE. If `map = TRUE` then map of source data is included in report. Default value is FALSE}

\item{locale}{character string. Currently English (`'EN'`) and Russian (`'RU'`) locales are supported for HTML output. PDF is always rendered in English}

\item{pdf}{TRUE/FALSE. If `pdf = TRUE` then ouput is rendered as a single PDF file. Otherwise the output is rendered as single HTML file. Default is FALSE}
}
\value{
Generates a new out working directory with detailed reports
}
\description{
Generate reports with detailed hydrograph analysis
}
