% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{gr_set_locale}
\alias{gr_set_locale}
\title{Set the language that is used for plotting}
\usage{
gr_set_locale(locale = "EN")
}
\arguments{
\item{locale}{Character string locale. Currently only English (\code{'EN'}) and Russian (\code{'RU'}) locales are supported. Defaults to \code{'EN'}.}
}
\description{
Set the language that is used for plotting
}
\examples{
library(grwat)

data(spas) # example Spas-Zagorye data is included with grwat package

# separate
sep = gr_separate(spas, params = gr_get_params(reg = 'Midplain'))

# Default is English
gr_plot_sep(sep, 1978) 

# Change to Russian
gr_set_locale('RU')
gr_plot_sep(sep, 1978) 

# Restore English
gr_set_locale('EN')
gr_plot_sep(sep, 1978) 

}
