% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{gr_ss}
\alias{gr_ss}
\title{Perform scale-space transform of hydrograph}
\usage{
gr_ss(df, year = NULL, sigma = NULL)
}
\arguments{
\item{df}{hydrograph data frame containing Date and Qin columns}

\item{year}{year to be analyzed (will be used to filter Date)}

\item{sigma}{maximum sigma used for scale-space smoothing}
}
\value{
Scale4C object with full scale-space transform information
}
\description{
Perform scale-space transform of hydrograph
}
