% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{gr_ss_tree}
\alias{gr_ss_tree}
\title{Get scale-space tree data frame from Scale4C object}
\usage{
gr_ss_tree(ssdata)
}
\arguments{
\item{ssdata}{}
}
\value{

}
\description{
Get scale-space tree data frame from Scale4C object
}
