% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{gr_summarize}
\alias{gr_summarize}
\title{Calculate various summary stats for separated hydrograph}
\usage{
gr_summarize(tab)
}
\arguments{
\item{tab}{data frame resulting from `separate()` function}
}
\value{
data frame with one row for each water-resources year and multiple columns of statistics
}
\description{
Calculate various summary stats for separated hydrograph
}
