% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_to_pardf}
\alias{gr_to_pardf}
\title{Convert list of parameters to data frame}
\usage{
gr_to_pardf(params)
}
\arguments{
\item{params}{\code{list} of \code{list}s of hydrograph separation parameters as returned in \code{params} attribute by \code{\link[=gr_separate]{gr_separate()}}  with \code{debug = TRUE}.}
}
\value{
\code{tibble} data frame with tabular representation of parameters
}
\description{
Convert list of parameters to data frame
}
