% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grw_animate}
\alias{grw_animate}
\title{Animate discharge through years}
\usage{
grw_animate(df, locale = "EN")
}
\arguments{
\item{df}{data.framewith discarge values in Qin variable}

\item{locale}{}
}
\value{

}
\description{
Animate discharge through years
}
