% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grw_get_variables}
\alias{grw_get_variables}
\title{Get hydrograph parameters list}
\usage{
grw_get_variables()
}
\value{
data.frame of parameters
}
\description{
Get hydrograph parameters list
}
\examples{
grwat::get_variables()
}
