% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{grw_map}
\alias{grw_map}
\title{Map reanalysis and basin data}
\usage{
grw_map(pts, pts_sel, region, buffer = NULL)
}
\arguments{
\item{pts}{sf object with reanalysis points}

\item{pts_sel}{sf object with selected reanalysis points}

\item{region}{sf object with basin region}

\item{buffer}{sf object with buffered basin region}
}
\value{
plot object
}
\description{
Map reanalysis and basin data
}
