% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grw_plot_sep}
\alias{grw_plot_sep}
\title{Plot hydrograph separation}
\usage{
grw_plot_sep(
  df,
  years = NULL,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{data.frame with hydrograph separation}

\item{years}{a vector of years to be plotted}

\item{layout}{layout matrix}

\item{pagebreak}{logical. Whether to break page between plots (needed for reporting). Defaults to FALSE}

\item{locale}{string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{
ggplot2 objects
}
\description{
Plot hydrograph separation
}
