% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grw_plot_tests}
\alias{grw_plot_tests}
\title{Plot change year density}
\usage{
grw_plot_tests(tests, locale = "EN")
}
\arguments{
\item{tests}{result of `test_variables()`}

\item{locale}{}
}
\value{
plots density plot for change year
}
\description{
Plot change year density
}
