% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{grw_read_vars}
\alias{grw_read_vars}
\title{Read file with hydrograph variables}
\usage{
grw_read_vars(file_tot)
}
\arguments{
\item{total_file}{Path to a file with hydrograph characteristics variables}
}
\value{
data.frame
}
\description{
Read file with hydrograph variables
}
