% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{grw_set_params}
\alias{grw_set_params}
\title{Change separation parameters through graphical interface}
\usage{
grw_set_params(params)
}
\arguments{
\item{params}{Parameters list}
}
\value{

}
\description{
Change separation parameters through graphical interface
}
