% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{join_reanalysis}
\alias{join_reanalysis}
\title{Join reanalysis data based on geographic location and time}
\usage{
join_reanalysis(wd, bufsize = 50000)
}
\arguments{
\item{wd}{Character. Working directory with specified structure}

\item{bufsize}{Size of a buffer that is used to select reanalysis data
around each basin}
}
\value{
Generates water level files enriched with reanalysis data.
  Produces text reports on a number of a processed gauges in each basin
}
\description{
Join reanalysis data based on geographic location and time
}
