% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{kable_tests}
\alias{kable_tests}
\title{Kable p-values table by coloring it using green-yellow-red palette}
\usage{
kable_tests(tests)
}
\arguments{
\item{tests}{Test result returned by run_tests}
}
\value{
kabled version of p-values table coloured
}
\description{
Kable p-values table by coloring it using green-yellow-red palette
}
