% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{An arbitrary number of ggplot objects. Listed by colon}

\item{plotlist}{A list of ggplot objects}

\item{cols}{Integer. Number of columns in layout.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored. 
If the layout is something like `matrix(c(1,2,3,3), nrow=2, byrow=TRUE)`, 
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}
}
\value{
Plots the passed ggplot objects in a specified matrix layout
}
\description{
Produces matrix layout for ggplot graphics. 
Taken from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
\dontrun{
multiplot(g1, g2, layout=matrix(c(1,2,3,3), nrow=2, byrow=TRUE))
}
}
