% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_longterm}
\alias{plot_longterm}
\title{Plot long-term characteristics for two periods}
\usage{
plot_longterm(df, ..., tests = NULL, locale = "EN")
}
\arguments{
\item{df}{}

\item{locale}{}

\item{year}{}
}
\description{
Plot long-term characteristics for two periods
}
