% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_parameters}
\alias{plot_parameters}
\title{Plot interannual parameters}
\usage{
plot_parameters(df, names = NULL, tests = NULL, locale = "EN")
}
\arguments{
\item{df}{data.frame produced by separation function}

\item{tests}{}

\item{locale}{}
}
\description{
Plot interannual parameters
}
