% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_periods}
\alias{plot_periods}
\title{Plot long-term characteristics for two periods}
\usage{
plot_periods(df, year = 1900, change_year = NULL, fixedyear = TRUE,
  locale = "EN")
}
\arguments{
\item{df}{}

\item{year}{}

\item{locale}{}
}
\description{
Plot long-term characteristics for two periods
}
