% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_separation}
\alias{plot_separation}
\title{Plot hydrograph separation}
\usage{
plot_separation(df, years = NULL, layout = as.matrix(1),
  pagebreak = FALSE, locale = "EN")
}
\arguments{
\item{df}{}

\item{years}{}

\item{layout}{}

\item{pagebreak}{}

\item{locale}{}
}
\description{
Plot hydrograph separation
}
