% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ss}
\alias{plot_ss}
\title{Plot scale-space tree of hydrograph}
\usage{
plot_ss(sstree, df = NULL, year = NULL, inverse = FALSE)
}
\arguments{
\item{sstree}{scale-space tree data frame resulting from `ss_tree()` function}

\item{df}{hydrograph data frame containing Date and Qin columns}

\item{year}{year to be analyzed (must be the same as used for sstree generation)}
}
\value{
ggplot2 object
}
\description{
Plot scale-space tree of hydrograph
}
