% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_variables}
\alias{plot_variables}
\title{Plot interannual parameters}
\usage{
plot_variables(
  df,
  ...,
  tests = NULL,
  exclude = NULL,
  smooth = TRUE,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{data.frame produced by separation function (read by `read_separation()`)}

\item{...}{quoted sequence of variable names}

\item{tests}{tests list for the same variables (generated by `test_variables()` function)}

\item{exclude}{integer vector. A vector of years to be excluded from plotting}

\item{smooth}{logical. If true then local smoothing regression is plotted. Defaults to TRUE.}

\item{layout}{layout matrix}

\item{pagebreak}{logical. Whether to break page between plots (needed for reporting). Defaults to FALSE}

\item{locale}{string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{

}
\description{
Plot interannual parameters
}
