% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process}
\alias{process}
\title{Process gauge hydrological data}
\usage{
process(hdata, rean, buffer)
}
\arguments{
\item{hdata}{Water discharge series. Data frame containing 4 columns: YYYY, MM, DD and water level}

\item{rean}{List. Contains three components: days series, values and location points}

\item{buffer}{Region to select reanalysis data}
}
\value{
List of two components: water level table enriched with reanalysis data
  and selected points
}
\description{
Process gauge hydrological data
}
\examples{
\dontrun{
grwat::process(hdata, rean, buffer)
}
}
