% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_interim}
\alias{read_interim}
\title{Read ERA-INTERIM reanalysis NetCDF files with temperature and precipitation}
\usage{
read_interim(file_prec, file_temp)
}
\arguments{
\item{file_prec}{Precipitation NetCDF file}

\item{file_temp}{Temperature NetCDF file}
}
\value{
List containing time series, precipitation series, temperature series 
  and spatial points (sf)
}
\description{
Read ERA-INTERIM reanalysis NetCDF files with temperature and precipitation
}
\examples{
\dontrun{
grwat::read_interim("rean/prec.nc", "rean/temp.nc")
}
}
