% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_separation}
\alias{read_separation}
\title{Read file with hydrograph separation}
\usage{
read_separation(file_sep)
}
\arguments{
\item{file_sep}{Path to a file with hydrograph separation data}
}
\value{
data.frame
}
\description{
Read file with hydrograph separation
}
