% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_total}
\alias{read_total}
\title{Read file with hydrograph parameters}
\usage{
read_total(file_tot)
}
\arguments{
\item{total_file}{Path to a file}
}
\value{
data.frame
}
\description{
Read file with hydrograph parameters
}
