% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_variables}
\alias{read_variables}
\title{Read file with hydrograph variables}
\usage{
read_variables(file_tot)
}
\arguments{
\item{total_file}{Path to a file}
}
\value{
data.frame
}
\description{
Read file with hydrograph variables
}
