% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_basins}
\alias{report_basins}
\title{Generate reports with detailed hydrograph analysis}
\usage{
report_basins(wd)
}
\arguments{
\item{wd}{Character. A path to a working directory with specified structure}
}
\value{
Generates a new out working directory with detailed reports
}
\description{
Generate reports with detailed hydrograph analysis
}
