% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_gauge}
\alias{report_gauge}
\title{Generate report for the specified gauge folder}
\usage{
report_gauge(wd)
}
\arguments{
\item{wd}{A path to a gauge directory}
}
\description{
Generate report for the specified gauge folder
}
