% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{run_tests}
\alias{run_tests}
\title{Run various tests on interannual characteristics}
\usage{
run_tests(df, fixedyear = FALSE, year = NULL, locale = "EN")
}
\arguments{
\item{df}{data.frame produced by separation function}

\item{fixedyear}{use fixed year value for long-term changes stimation. Default is FALSE}

\item{year}{fixed year value for long-term changes estimation}

\item{locale}{locale for p-value table}
}
\value{
List of testing results
}
\description{
Run various tests on interannual characteristics
}
