% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{set_separation_params}
\alias{set_separation_params}
\title{Change separation parameters through graphical interface}
\usage{
set_separation_params(params)
}
\arguments{
\item{params}{Parameters list}
}
\value{

}
\description{
Change separation parameters through graphical interface
}
