% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_transform_opt}
\alias{st_transform_opt}
\title{Project data to the best conic projection}
\usage{
st_transform_opt(s)
}
\arguments{
\item{s}{`sf` (simple features) object}
}
\value{
`sf` (simple features) object projected to automatically selected projection
}
\description{
Project data to the best conic projection
}
