% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{test_variables}
\alias{test_variables}
\title{Run various tests on interannual characteristics}
\usage{
test_variables(df, ..., change_year = NULL, locale = "EN")
}
\arguments{
\item{df}{data.frame produced by separation function}

\item{...}{fields (quoted)}

\item{locale}{locale for p-value table}

\item{fixedyear}{use fixed year value for long-term changes stimation. Default is FALSE}

\item{year}{fixed year value for long-term changes estimation}
}
\value{
List of testing results
}
\description{
Run various tests on interannual characteristics
}
