% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{update_core}
\alias{update_core}
\title{This is a technical function to update grwat core from a separate project}
\usage{
update_core()
}
\value{

}
\description{
This is a technical function to update grwat core from a separate project
}
