% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_animate}
\alias{gr_animate}
\title{Animate discharge through years}
\usage{
gr_animate(
  df,
  plot = TRUE,
  file = NULL,
  fps = 20,
  kframes = 10,
  width = 800,
  height = 600,
  locale = "EN"
)
}
\arguments{
\item{df}{A data.frame with date (1st) and discharge (2nd) columns.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{
The return value of the \code{\link[gganimate:renderer]{gganimate::renderer()}} function
}
\description{
Animate discharge through years
}
