% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_animate}
\alias{gr_animate}
\title{Animate hydrograph}
\usage{
gr_animate(
  df,
  plot = TRUE,
  file = NULL,
  fps = 20,
  kframes = 10,
  width = 800,
  height = 600
)
}
\arguments{
\item{df}{\code{data.frame} with date (1st) and discharge (2nd) columns.}

\item{plot}{Boolean. Plot the animation in the viewer? Defaults to \code{TRUE}.}

\item{file}{Character string path to the output file.}

\item{fps}{Integer value of the frames per second. Defaults to \code{20}.}

\item{kframes}{Integer value setting how many frames are interpolated between the year. Defaults to \code{10}, which means that if \code{fps = 20}, then two years are passed each second.}

\item{width}{Integer width of the animation (in pixels). Defaults to 800.}

\item{height}{Integer height of the animation (in pixels). Defaults to 600}
}
\value{
The return value of the \code{\link[gganimate:animate]{gganimate::animate()}} function
}
\description{
A convenient wrapper around \code{\link[gganimate:animate]{gganimate::animate()}} to animate river hydrograph through years.
}
\examples{
if (require("gganimate") && require("transformr") && require("gifski") && require("png")) {
  
  library(grwat)
  
  # example Spas-Zagorye data is included with grwat package
  data(spas)
  head(spas)
  
  # select years if needed
  df = spas[spas$Date < as.Date('1960-01-01'), ]
  
  # animate
  gr_animate(df)
  
}
}
