% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_baseflow}
\alias{gr_baseflow}
\title{Extract baseflow}
\usage{
gr_baseflow(
  Q,
  a = 0.925,
  k = 0.975,
  C = 0.05,
  aq = -0.5,
  passes = 3,
  padding = 30,
  method = "lynehollick"
)
}
\arguments{
\item{Q}{Numeric runoff vector.}

\item{a}{Numeric value of a filtering parameter used in \code{'chapman'}, \code{'jakeman'} and \code{'lynehollick'} methods. Defaults to \code{0.925}.}

\item{k}{Numeric value of a filtering parameter used in \code{'boughton'} and \code{'maxwell'} methods. Defaults to \code{0.975}.}

\item{C}{Numeric value of a separation shape parameter used in \code{'boughton'}, \code{'jakeman'} and \code{'maxwell'} methods}

\item{aq}{Numeric value of a filtering parameter used in \code{'jakeman'} method. Defaults to \code{-0.5}.}

\item{passes}{Integer number of filtering iterations. The first iteration is forward, second is backward, third is forward and so on. Defaults to \code{3}.}

\item{padding}{Integer number of elements padded at the beginning and ending of runoff vector to reduce boundary effects. Defaults to \code{30}.}

\item{method}{Character string to set baseflow filtering method. Available methods are \code{'boughton'}, \code{'chapman'}, \code{'furey'}, \code{'jakeman'}, \code{'kudelin'}, \code{'lynehollick'} and \code{'maxwell'}. Default is \code{'lynehollick'}, which corresponds to Lyne-Hollick (1979) hydrograph separation method.}
}
\value{
Numeric baseflow vector
}
\description{
Extract baseflow from hydrological series using the filtering approach
}
\examples{

}
