% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{gr_buffer_geo}
\alias{gr_buffer_geo}
\title{Buffer sf objects in geographic coordinates through conic projection}
\usage{
gr_buffer_geo(g, bufsize)
}
\arguments{
\item{g}{\code{sf} or \code{sfg} object in geographic coordinates}

\item{bufsize}{double. Size of a buffer, in metres}
}
\value{
\code{sf} or \code{sfg} object, buffered to \code{bufsize} distance
}
\description{
Buffer sf objects in geographic coordinates through conic projection
}
\examples{
\dontrun{
gr_buffer_geo(basin, 50000)
}
}
