% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_check_data}
\alias{gr_check_data}
\title{Check the correctness of data frame for separating}
\usage{
gr_check_data(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Check the correctness of data frame for separating
}
\examples{

}
