% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{gr_get_gaps}
\alias{gr_get_gaps}
\title{Get gaps in the data}
\usage{
gr_get_gaps(hdata)
}
\arguments{
\item{hdata}{a data frame where the first column is date, and the second column is discharge value}
}
\value{
a data frame with periods of data and periods of gaps
}
\description{
Get gaps in the data
}
