% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_help_params}
\alias{gr_help_params}
\title{Get the information about parameters used to separate the hydrograph}
\usage{
gr_help_params()
}
\value{
a table with parameter names
}
\description{
Get the information about parameters used to separate the hydrograph
}
