% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{gr_help_vars}
\alias{gr_help_vars}
\title{Hydrograph separation variables}
\usage{
gr_help_vars()
}
\value{
\code{data.frame} of hydrograph separation variables
}
\description{
Use this function to learn the meaning of the variables that are calculated by \code{\link[=gr_summarize]{gr_summarize()}}.
}
\examples{
library(grwat)

gr_help_vars()
}
