% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{gr_kable_tests}
\alias{gr_kable_tests}
\title{Kable p-values table by coloring it using green-yellow-red palette}
\usage{
gr_kable_tests(tests, locale = "EN", format = "latex")
}
\arguments{
\item{tests}{Test result returned by `test_variables()` function}

\item{locale}{character string. Currently English (`'EN'`) and Russian (`'RU'`) locales are supported for HTML output. PDF is always rendered in English}

\item{format}{character string. Currently `'latex'` or `'html'` are supported}
}
\value{
kabled version of p-values table coloured
}
\description{
Kable p-values table by coloring it using green-yellow-red palette
}
