% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_acf}
\alias{gr_plot_acf}
\title{Plot the autocorrelation function (ACF) for hydrograph time series}
\usage{
gr_plot_acf(hdata, autocorr = 0.7, max_lag = 30)
}
\arguments{
\item{hdata}{\code{data.frame} with two columns: date and discharge}

\item{autocorr}{Numeric value of the autocorrelation for which the time period will be highlighted.}
}
\value{
ACF plot
}
\description{
Plot the autocorrelation function (ACF) for hydrograph time series
}
