% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_hori}
\alias{gr_plot_hori}
\title{A convenient wrapper around \code{\link[ggHoriPlot:geom_horizon]{ggHoriPlot::geom_horizon()}} to visualize multiple river hydrographs at once}
\usage{
gr_plot_hori(df, years, pal = "Blues", rev = T, scale = 6, locale = "EN")
}
\arguments{
\item{df}{\code{data.frame} with date (1st) and discharge (2nd) columns.}

\item{years}{Integer vector of years to be plotted.}

\item{pal}{Numeric or character string. Color palette identifier passed to \code{\link[ggplot2:scale_brewer]{ggplot2::scale_fill_distiller()}}.}

\item{rev}{Boolean. Reverse the palette? Defaults to \code{FALSE}.}

\item{scale}{Numeric scale factor passed to \code{\link[ggHoriPlot:geom_horizon]{ggHoriPlot::geom_horizon()}}. Defaults to \code{0.01}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\value{
A \code{ggplot2} object
}
\description{
A convenient wrapper around \code{\link[ggHoriPlot:geom_horizon]{ggHoriPlot::geom_horizon()}} to visualize multiple river hydrographs at once
}
