% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_matrix}
\alias{gr_plot_matrix}
\title{Plot discharge value, source and season via raster-based visualization through years}
\usage{
gr_plot_matrix(df, years = NULL, type = "value", locale = "EN")
}
\arguments{
\item{df}{Hydrograph separation data.frame produced by \code{gr_separate()}.}

\item{years}{Integer vector of years to be plotted.}

\item{type}{String. Supported options are \code{'value'}, \code{'components'}, and \code{'season'}. Defaults to \code{'value'}}

\item{locale}{String locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot discharge value, source and season via raster-based visualization through years
}
