% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_matrix}
\alias{gr_plot_matrix}
\title{Plot discharge value, source and season via raster-based visualization through years}
\usage{
gr_plot_matrix(df, years = NULL, type = "value", locale = "EN")
}
\arguments{
\item{df}{\code{data.frame} of hydrograph separation  produced by \code{\link[=gr_separate]{gr_separate()}}.}

\item{years}{Integer vector of years to be plotted.}

\item{type}{Character string. Supported options are 'value', 'components', and 'season'. Defaults to 'value'.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot discharge value, source and season via raster-based visualization through years
}
