% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_minmonth}
\alias{gr_plot_minmonth}
\title{Plot minimum discharge month}
\usage{
gr_plot_minmonth(
  df,
  year = NULL,
  exclude = NULL,
  tests = NULL,
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{\code{data.frame} of hydrograph and meteorological variables as produced by \code{\link[=gr_summarize]{gr_summarize()}}.}

\item{year}{Integer. Change year value to separate two periods.}

\item{exclude}{Integer vector of years to be excluded from plotting.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\description{
Generate a histogram of a minimum discharge month for two periods: before and after the change year set by \code{year} parameter.
}
