% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_periods}
\alias{gr_plot_periods}
\title{Plot long-term changes of hydrograph variables for two periods}
\usage{
gr_plot_periods(
  df,
  ...,
  year = NULL,
  exclude = NULL,
  tests = NULL,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{data.frame produced by description function (read by `read_variables()`)}

\item{...}{quoted sequence of variable names}

\item{year}{change year value to separate two periods (overridden by tests if it is supplied)}

\item{exclude}{integer vector. A vector of years to be excluded from plotting}

\item{tests}{tests list for the same variables (generated by `test_variables()` function)}

\item{layout}{layout matrix}

\item{pagebreak}{logical. Whether to break page between plots (needed for reporting). Defaults to FALSE}

\item{locale}{string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{
ggplot2 objects
}
\description{
Plot long-term changes of hydrograph variables for two periods
}
