% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_periods}
\alias{gr_plot_periods}
\title{Plot long-term hydrograph variable changes}
\usage{
gr_plot_periods(
  df,
  ...,
  year = NULL,
  exclude = NULL,
  tests = NULL,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{\code{data.frame} of hydrograph and meteorological variables produced by \code{\link[=gr_summarize]{gr_summarize()}}.}

\item{...}{Quoted sequence of variable names.}

\item{year}{Integer. Change year value to separate two periods (overridden by tests if it is supplied).}

\item{exclude}{Integer vector of years to be excluded from plotting.}

\item{tests}{Tests list for the same variables (generated by \code{\link[=test_variables]{test_variables()}} function)}

\item{layout}{\code{matrix} that encodes the order of plotting.}

\item{pagebreak}{Logical. Whether to break page between plots (needed for reporting). Defaults to \code{FALSE}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\description{
This function generates boxplots of the hydrograph separation variables produced by \code{\link[=gr_summarize]{gr_summarize()}}. The data for each variable is divided into two samples: before and after the change year either set by \code{year} parameter or extracted from \code{tests} (statistically estimated). Different background fill colors are used to differentiate seasons types.
}
