% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_ridge}
\alias{gr_plot_ridge}
\title{Ridgeline hydrograph plot}
\usage{
gr_plot_ridge(
  df,
  years,
  pal = 4,
  rev = FALSE,
  scale = 0.01,
  alpha = 0.8,
  locale = "EN"
)
}
\arguments{
\item{df}{\code{data.frame} with date (1st) and discharge (2nd) columns.}

\item{years}{Integer vector of years to be plotted.}

\item{pal}{Numeric or character string. Color palette identifier passed to \code{\link[ggplot2:scale_brewer]{ggplot2::scale_fill_distiller()}}.}

\item{rev}{Boolean. Reverse the palette? Defaults to \code{FALSE}.}

\item{scale}{Numeric scale factor passed to \code{\link[ggridges:geom_ridgeline]{ggridges::geom_ridgeline()}}. Defaults to \code{0.01}.}

\item{alpha}{Numeric opacity value of the ridgeline plot. Defaults to \code{0.8}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\description{
A convenient wrapper around \code{\link[ggridges:geom_ridgeline]{ggridges::geom_ridgeline()}} to visualize multiple river hydrographs at once.
}
\examples{

}
