% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_sep}
\alias{gr_plot_sep}
\title{Plot hydrograph separation}
\usage{
gr_plot_sep(
  df,
  years = NULL,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN",
  prec = FALSE,
  temp = FALSE,
  span = 5
)
}
\arguments{
\item{df}{Hydrograph separation data.frame produced by \code{gr_separate()}.}

\item{years}{Integer vector of years to be plotted.}

\item{layout}{Layout matrix.}

\item{pagebreak}{Logical. Whether to break page between plots (needed for reporting). Defaults to FALSE.}

\item{locale}{String locale. Currently only English locale is supported. Defaults to 'EN'.}

\item{prec}{Boolean. Add precipitation curve to the plot? Defaults to FALSE.}

\item{temp}{Boolean. Add temperature curve to the plot? Defaults to FALSE.}
}
\value{
ggplot2 objects
}
\description{
Plot hydrograph separation
}
