% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_tests}
\alias{gr_plot_tests}
\title{Plot change year density}
\usage{
gr_plot_tests(tests, locale = "EN")
}
\arguments{
\item{tests}{\code{list} of tests generated by \code{\link[=test_variables]{test_variables()}}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\description{
The function extracts change years from results of \code{\link[=test_variables]{test_variables()}} and plots their probability density. Since for every variable the change year is individual, this procedure allows finding the one most probable year, which is the mode of the distribution. This year is highlighted by the line and labeled on the plot.
}
