% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_vars}
\alias{gr_plot_vars}
\title{Plot interannual parameters}
\usage{
gr_plot_vars(
  df,
  ...,
  tests = NULL,
  exclude = NULL,
  smooth = TRUE,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{\code{data.frame} of hydrograph and meteorological variables produced by \code{\link[=gr_summarize]{gr_summarize()}}.}

\item{...}{Quoted sequence of variable names.}

\item{tests}{\code{list} of tests for the same variables (generated by \code{\link[=test_variables]{test_variables()}} function).}

\item{exclude}{Integer vector of years to be excluded from plotting.}

\item{smooth}{Logical. If \code{TRUE} then local smoothing regression is plotted. Defaults to \code{TRUE}.}

\item{layout}{\code{matrix} that encodes the order of plotting.}

\item{pagebreak}{Logical. Whether to break page between plots (\code{\link[=gr_report]{gr_report()}}). Defaults to \code{FALSE}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{

}
\description{
Plot interannual parameters
}
