% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_vars}
\alias{gr_plot_vars}
\title{Plot interannual parameters}
\usage{
gr_plot_vars(
  df,
  ...,
  tests = NULL,
  exclude = NULL,
  smooth = TRUE,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{Hydrograph and meteorological variables data.frame produced by \code{gr_summarize()}.}

\item{...}{Quoted sequence of variable names.}

\item{tests}{Tests list for the same variables (generated by \code{test_variables()} function).}

\item{exclude}{Integer vector of years to be excluded from plotting.}

\item{smooth}{Logical. If true then local smoothing regression is plotted. Defaults to TRUE.}

\item{layout}{Layout matrix.}

\item{pagebreak}{Logical. Whether to break page between plots (needed for reporting). Defaults to FALSE.}

\item{locale}{String locale. Currently only English locale is supported. Defaults to 'EN'.}
}
\value{

}
\description{
Plot interannual parameters
}
