% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr_plot_vars}
\alias{gr_plot_vars}
\title{Plot interannual hydrograph variable changes}
\usage{
gr_plot_vars(
  df,
  ...,
  tests = NULL,
  exclude = NULL,
  smooth = TRUE,
  layout = as.matrix(1),
  pagebreak = FALSE,
  locale = "EN"
)
}
\arguments{
\item{df}{\code{data.frame} of hydrograph and meteorological variables produced by \code{\link[=gr_summarize]{gr_summarize()}}.}

\item{...}{Quoted sequence of variable names.}

\item{tests}{\code{list} of tests for the same variables (generated by \code{\link[=test_variables]{test_variables()}} function). If tests are specified, then they are added to the plot.}

\item{exclude}{Integer vector of years to be excluded from plotting.}

\item{smooth}{Logical. If \code{TRUE} then local smoothing regression is plotted. Defaults to \code{TRUE}.}

\item{layout}{\code{matrix} that encodes the order of plotting.}

\item{pagebreak}{Logical. Whether to break page between plots (\code{\link[=gr_report]{gr_report()}}). Defaults to \code{FALSE}.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\description{
This function plots the hydrograph separation variables produced by \code{\link[=gr_summarize]{gr_summarize()}}. Different background fill colors and line types are used to differentiate seasons and variable types.
}
\examples{
library(grwat)

data(spas) # example Spas-Zagorye data is included with grwat package

# separate
sep = gr_separate(spas, params = gr_get_params(reg = 'Midplain'))

# summarize
vars = gr_summarize(sep)

# plot one selected variable
gr_plot_vars(vars, Qygr) 

# plot one selected variable
gr_plot_vars(vars, datestart) 

# plot two variables sequentially
gr_plot_vars(vars, date10w1, Wpol3)

# plot four variables in matrix layout
gr_plot_vars(vars, Qmax, Qygr, date10w1, Wpol3,
             layout = matrix(1:4, nrow = 2, byrow = TRUE)) 

# add tests calculated on the fly (only plotted variables are tested)
gr_plot_vars(vars, Qmax, Qygr, date10w1, Wpol3,
             layout = matrix(1:4, nrow = 2, byrow = TRUE),
             tests = TRUE) 

# calculate tests beforehand
tests = gr_test_vars(vars)
gr_plot_vars(vars, date10w1, Wpol3, DaysThawWin, Qmaxpavs,
             layout = matrix(1:4, nrow = 2, byrow = TRUE),
             tests = tests) 
}
