% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_separate}
\alias{gr_separate}
\title{Separates river hydrograph}
\usage{
gr_separate(df, params = grw_get_params(), alpha = 0.925, niter = 100)
}
\arguments{
\item{df}{data frame with six columns: year, month, date, discharge, temperature, precipitation}

\item{params}{list of separation parameters}

\item{order}{order, in which the columns are placed in data frame. Defaults to 'dmyqtp'}
}
\value{

}
\description{
Separates river hydrograph
}
