% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_separate}
\alias{gr_separate}
\title{Separates river hydrograph}
\usage{
gr_separate(df, params = gr_get_params(), alpha = 0.925, niter = 100)
}
\arguments{
\item{df}{data frame with four columns: date, discharge, temperature, precipitation}

\item{params}{list of separation parameters, as returned by \link{gr_get_params()} function}

\item{alpha}{smoothing parameter for groundwater separation algorithm, defaults to 0.925 for Lyne-Hollick method}

\item{niter}{number of iterations}
}
\value{

}
\description{
Separates river hydrograph
}
