% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{gr_set_param}
\alias{gr_set_param}
\title{Set the value of selected parameter for selected years in parameter list}
\usage{
gr_set_param(params, p, value, years = NULL)
}
\arguments{
\item{params}{\code{list} of \code{list}s of hydrograph separation parameters as returned in \code{params} attribute by \code{\link[=gr_separate]{gr_separate()}}  with \code{debug = TRUE}.}

\item{p}{Name of the parameter.}

\item{value}{Numeric value to set.}

\item{years}{Integer vector of years to modify. Defaults to \code{NULL}, which means that all years will be modified.}
}
\value{
\code{list} of \code{list}s — a modified version of \code{params}
}
\description{
Set the value of selected parameter for selected years in parameter list
}
