% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{gr_test_vars}
\alias{gr_test_vars}
\title{Test hydrograph changes}
\usage{
gr_test_vars(df, ..., year = NULL, exclude = NULL, locale = "EN")
}
\arguments{
\item{df}{\code{data.frame} as produced by \code{\link[=gr_summarize]{gr_summarize()}} function.}

\item{...}{Names of the tested variables (quoted).}

\item{year}{Integer value of year used to divide series in two samples compared by Student and Fisher tests. Defaults to \code{NULL} which means that the year is calculated automatically by Pettitt test.}

\item{exclude}{Integer vector of years to be excluded from tests.}

\item{locale}{Character string locale. Currently only English locale is supported. Defaults to \code{'EN'}.}
}
\value{
\code{list} of testing results
}
\description{
Use this function to test interannual changes or hydrograph separation variables returned by \code{\link[=gr_summarize]{gr_summarize()}}. Pettitt test is used to detect the change year — i.e. the year which divides the time series into the statistically most differing samples. Student (Welch) and Fisher tests are used to estimate the significance of mean and variance differences of these samples. Theil-Sen test calculates the trend slope value. Mann-Kendall test is performed to reveal the significance of the trend.
}
\details{
Number of observations formally required for various tests: Pettitt > 0, Mann-Kendall > 2, Theil-Sen > 1, Student > 1, Fisher > 1.
}
\examples{
library(grwat)

data(spas) # example Spas-Zagorye data is included with grwat package

# separate
sep = gr_separate(spas, params = gr_get_params(reg = 'Midplain'))

# summarize
vars = gr_summarize(sep)

# test all variables
tests = gr_test_vars(vars)

# view Pettitt test for Qygr
tests$ptt$Qygr

# view Fisher test for Q30s
tests$ft$Q30s

# test only Qygr and Q30s using 1978 as fixed year and excluding 1988-1991 yrs
gr_test_vars(vars, Qygr, Q30s, year = 1978, exclude = 1988:1991)
}
