% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_minmonth}
\alias{plot_minmonth}
\title{Plot histogram of minimum discharge month for two periods}
\usage{
plot_minmonth(df, change_year = NULL, pagebreak = FALSE, locale = "EN")
}
\arguments{
\item{df}{}

\item{change_year}{}

\item{locale}{}
}
\description{
Plot histogram of minimum discharge month for two periods
}
