% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_periods}
\alias{plot_periods}
\title{Plot long-term characteristics for two periods}
\usage{
plot_periods(df, ..., year = NULL, tests = NULL, layout = as.matrix(1),
  pagebreak = FALSE, locale = "EN")
}
\arguments{
\item{df}{}

\item{...}{}

\item{year}{}

\item{tests}{}

\item{layout}{}

\item{pagebreak}{}

\item{locale}{}
}
\description{
Plot long-term characteristics for two periods
}
