% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_variables}
\alias{plot_variables}
\title{Plot interannual parameters}
\usage{
plot_variables(df, ..., tests = NULL, smooth = TRUE,
  layout = as.matrix(1), pagebreak = FALSE, locale = "EN")
}
\arguments{
\item{df}{data.frame produced by separation function}

\item{tests}{}

\item{locale}{}
}
\description{
Plot interannual parameters
}
