% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_basins}
\alias{process_basins}
\title{Join reanalysis data based on geographic location and time}
\usage{
process_basins(wd, rean, bufsize = 50000, eol = "\\n", clear = TRUE)
}
\arguments{
\item{wd}{Path to a working directory with specified structure}

\item{rean}{Reanalysis data (as read by `read_interim()`)}

\item{bufsize}{Size of a buffer that is used to select reanalysis data
around each basin}

\item{eol}{The symbol used to specify end of line. Defaults to "\\n", which will work well on most of the systems. On some Windows setups, an alternative "\\r\\n" must be used if the standard option does not work.}

\item{clear}{Boolean. Whether to remove out directory before processing. Defaults to TRUE}
}
\value{
Generates water level files enriched with reanalysis data.
  Produces text reports on a number of a processed gauges in each basin
}
\description{
Join reanalysis data based on geographic location and time
}
\examples{
\dontrun{
grwat::process_basins(wd, rean, bufsize=50000, clear=TRUE)
}
}
