% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_gauge}
\alias{process_gauge}
\title{Process gauge directory}
\usage{
process_gauge(wd, rean, bufsize = 50000, eol = "\\n", filename = "in.txt")
}
\arguments{
\item{wd}{Path to a directory containing water level series data and basin region}

\item{rean}{List with reanalysis data (generated by read_interim function)}

\item{bufsize}{Buffer distance to select additional points around basin region}

\item{eol}{The symbol used to specify end of line. Defaults to "\\n", which will work well on most of the systems. On some Windows setups, an alternative "\\r\\n" must be used if the standard option does not work.}

\item{filename}{name of the hydro file. Defaults to `in.txt`}
}
\value{
Number of reanalysis points processed for buffered basin. All other
  operations are performed inside a directory
}
\description{
Process gauge directory
}
\examples{
\dontrun{
grwat::process_gauge(wd, rean, 50000)
}
}
