% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_separation}
\alias{read_separation}
\title{Read file with separation hydrograph}
\usage{
read_separation(file_sep)
}
\arguments{
\item{file_sep}{Path to a file}
}
\value{
data.frame
}
\description{
Read file with separation hydrograph
}
