% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{separate}
\alias{separate}
\title{Separates river hydrograph}
\usage{
separate(df, params)
}
\arguments{
\item{df}{data frame with six columns: year, month, date, discharge, temperature, precipitation}

\item{params}{list of separation parameters}
}
\value{

}
\description{
Separates river hydrograph
}
