% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_variables}
\alias{test_variables}
\title{Run various tests on interannual characteristics.
Required number of observations for various tests: 
Pettitt > 0, Mann.Kendall > 2, Theil-Sen slope > 1, Student > 1,}
\usage{
test_variables(df, ..., year = NULL, exclude = NULL, locale = "EN")
}
\arguments{
\item{df}{data.frame produced by separation function}

\item{...}{fields (quoted)}

\item{year}{fixed year value for long-term changes estimation}

\item{exclude}{integer vector. A vector of years to be excluded from tests}

\item{locale}{locale for p-value table}
}
\value{
List of testing results
}
\description{
Run various tests on interannual characteristics.
Required number of observations for various tests: 
Pettitt > 0, Mann.Kendall > 2, Theil-Sen slope > 1, Student > 1,
}
